Fix problem with duplicated player names on match create
Added virtual attribute to know if player is selected from autocomplete
(ie. it's an existing record) or a new player record must be created. So
far, since players are fetched by name on matches#create, it was not
possible to create a new player with the same name as an existing one.